requirements.txt
fastapi
uvicorn
deepface
opencv-python
ffmpeg-python
openai-whisper
python-multipart
pydantic
numpy<2
pandas<2.2.2
pyarrow<15
tensorflow-macos>=2.13.0
tensorflow-metal>=1.0.0
tf-keras 